use crates/data_components/src/duckdb/federation.rs
