+++ crates/data_components/src/duckdb/federation.rs
@@ -1,0 +1,46 @@
+/*
+Copyright 2024 The Spice.ai OSS Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+     https://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+use datafusion_federation::{FederatedTableProviderAdaptor, FederatedTableSource};
+use datafusion_federation_sql::{SQLFederationProvider, SQLTableSource};
+use std::sync::Arc;
+
+use super::sql_table::DuckDBTable;
+use datafusion::{datasource::TableProvider, error::Result as DataFusionResult};
+
+impl<T, P> DuckDBTable<T, P> {
+    fn create_federated_table_source(
+        self: Arc<Self>,
+    ) -> DataFusionResult<Arc<dyn FederatedTableSource>> {
+        let table_name = self.base_table.table_reference.to_quoted_string();
+        let schema = Arc::clone(&Arc::clone(&self).base_table.schema());
+        let fed_provider = Arc::new(SQLFederationProvider::new(self));
+        Ok(Arc::new(SQLTableSource::new_with_schema(
+            fed_provider,
+            table_name,
+            schema,
+        )?))
+    }
+
+    pub fn create_federated_table_provider(
+        self: Arc<Self>,
+    ) -> DataFusionResult<FederatedTableProviderAdaptor> {
+        let table_source = Self::create_federated_table_source(Arc::clone(&self))?;
+        Ok(FederatedTableProviderAdaptor::new_with_provider(
+            table_source,
+            self,
+        ))
+    }
+}
